fix loi cart, them thuoc tinh amount cho Cart

+++ backend/src/main/java/com/example/shop/payloads/dto/CartItemDTO.java
@@ -17,9 +17,9 @@ public class CartItemDTO {
 
     private String imageUrl;
 
-    private Double price;
+    private Long price;
 
     private Integer quantity;
 
-    private Double subTotal;
+    private Long subTotal;
 }
